Added placeholder option
Add placeholder option

Added a `placeholder` attribute to the underlying textarea and listed it as a prop on the `Codemirror` component.

The component's `didComponentMount` method now checks to see whether the placeholder prop is present before calling `setValue` with the initial value. If placeholder is present, it sets the value as blank, thereby allowing the placeholder to show up.

Updated the example accordingly.